+++ include/granit/renderer/swapchain.hpp
@@ -182,7 +182,7 @@ public:
       return result::success;
     }
     const auto value = granit_swapchain_destroy(renderer_, handle_);
-    if (value == GRANIT_SUCCESS) {
+    if (value == GRANIT_SUCCESS || value == GRANIT_ERROR_INVALID_HANDLE) {
       renderer_ = GRANIT_NULL_HANDLE;
       handle_ = GRANIT_NULL_HANDLE;
     }