+++ include/granit/window/window.hpp
@@ -65,7 +65,7 @@ public:
     if (!valid())
       return result::success;
     const auto value = granit_window_system_destroy(handle_);
-    if (value == GRANIT_SUCCESS)
+    if (value == GRANIT_SUCCESS || value == GRANIT_ERROR_INVALID_HANDLE)
       handle_ = GRANIT_NULL_HANDLE;
     return from_native(value);
   }
@@ -115,7 +115,7 @@ public:
     if (!valid())
       return result::success;
     const auto value = granit_window_destroy(system_, handle_);
-    if (value == GRANIT_SUCCESS) {
+    if (value == GRANIT_SUCCESS || value == GRANIT_ERROR_INVALID_HANDLE) {
       system_ = GRANIT_NULL_HANDLE;
       handle_ = GRANIT_NULL_HANDLE;
     }